refactor: rename `uefi_wrapper` to `uefi`
Too long name.
