+++ Cargo.toml
@@ -7,7 +7,7 @@ members = [
     "libs/frame_allocator",
     "libs/kernel_mmap",
     "libs/qemu",
-    "libs/uefi_wrapper",
+    "libs/uefi",
 ]
 
 [profile.dev]
