+++ bootx64/Cargo.toml
@@ -25,7 +25,7 @@ qemu = { path = "../libs/qemu", optional = true }
 qemu_print = { version = "0.1.0", features = ["stable"], default-features = false }
 rlibc = "1.0.0"
 spinning_top = "0.2.4"
-uefi_wrapper = { path = "../libs/uefi_wrapper" }
+uefi = { path = "../libs/uefi" }
 x86_64 = { version = "0.14.4", features = ["instructions", "external_asm"], default-features = false }
 
 [build-dependencies]
