+++ libs/boot_info/Cargo.toml
@@ -6,5 +6,5 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 aligned_ptr = "0.1.0"
-uefi_wrapper = { path = "../uefi_wrapper" }
+uefi = { path = "../uefi" }
 x86_64 = { version = "0.14.4", default-features = false }
