+++ libs/boot_info/src/lib.rs
@@ -2,7 +2,7 @@
 
 use {
     aligned_ptr::slice,
-    uefi_wrapper::service::boot::{
+    uefi::service::boot::{
         MemoryDescriptor, MemoryType, ACPI_MEMORY_NVS, ACPI_RECLAIM_MEMORY, BOOT_SERVICES_CODE,
         BOOT_SERVICES_DATA, CONVENTIONAL_MEMORY, LOADER_CODE, LOADER_DATA, MEMORY_MAPPED_IO,
         MEMORY_MAPPED_IO_PORT_SPACE, MEMORY_MORE_RELIABLE, MEMORY_NV, MEMORY_RO, MEMORY_RP,
