+++ libs/frame_allocator/Cargo.toml
@@ -9,5 +9,5 @@ license = "MIT OR Apache-2.0"
 arrayvec = { version = "0.7.1", default-features = false }
 bit_field = "0.10.1"
 os_units = "0.4.0"
-uefi_wrapper = { path = "../uefi_wrapper" }
+uefi = { path = "../uefi" }
 x86_64 = { version = "0.14.4", default-features = false }
