+++ libs/uefi/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "uefi_wrapper"
+name = "uefi"
 version = "0.1.0"
 authors = ["Hiroki Tokunaga <tokusan441@gmail.com>"]
 edition = "2018"
