+++ libs/uefi/src/error.rs
