+++ libs/uefi/src/protocols/console/edid.rs
