+++ libs/uefi/src/protocols/console/graphics_output.rs
