+++ libs/uefi/src/protocols/console/mod.rs
