+++ libs/uefi/src/protocols/console/simple_text_output.rs
