+++ libs/uefi/src/result.rs
